+++ src/error.rs
@@ -35,7 +35,7 @@ pub enum ConnectionError {
     /// A pseudo-header is not valid in this message.
     #[error("Unexpected pseudo header {0:?}")]
     UnexpectedPseudo(&'static str),
-    /// A WINDOW_UPDATE increment was zero.
+    /// A `WINDOW_UPDATE` increment was zero.
     #[error("Window update value is zero")]
     ZeroWindowUpdateValue,
     /// A flow-control window overflowed.
@@ -53,6 +53,9 @@ pub enum ConnectionError {
     /// Frame reading timed out.
     #[error("Read timeout")]
     ReadTimeout,
+    /// Frame write timed out.
+    #[error("Write timeout")]
+    WriteTimeout,
 }
 
 impl ConnectionError {
@@ -99,6 +102,9 @@ impl ConnectionError {
                 GoAway::new(Reason::NO_ERROR).set_data("Keep-alive timeout")
             }
             ConnectionError::ReadTimeout => GoAway::new(Reason::NO_ERROR).set_data("Frame read timeout"),
+            ConnectionError::WriteTimeout => {
+                GoAway::new(Reason::NO_ERROR).set_data("Frame write timeout")
+            }
         }
     }
 }
@@ -121,6 +127,7 @@ impl ErrorDiagnostic for ConnectionError {
             ConnectionError::StreamResetsLimit => "h2-conn-StreamResetsLimit",
             ConnectionError::KeepaliveTimeout => "h2-conn-KeepaliveTimeout",
             ConnectionError::ReadTimeout => "h2-conn-ReadTimeout",
+            ConnectionError::WriteTimeout => "h2-conn-WriteTimeout",
         }
     }
 }
@@ -154,10 +161,10 @@ pub enum StreamError {
     /// The stream flow-control window overflowed.
     #[error("Window value is overflowed")]
     WindowOverflowed,
-    /// A stream WINDOW_UPDATE increment was zero.
+    /// A stream `WINDOW_UPDATE` increment was zero.
     #[error("Zero value for window")]
     WindowZeroUpdateValue,
-    /// Trailers were received without END_STREAM.
+    /// Trailers were received without `END_STREAM`.
     #[error("Trailers headers without end of stream flags")]
     TrailersWithoutEos,
     /// The content-length header is invalid.